+++ R/plot.binned_residuals.R
@@ -1,7 +1,7 @@
 #' @export
 plot.see_binned_residuals <- function(
   x,
-  linewidth = 0.7,
+  size_line = 0.7,
   size_point = 2.2,
   size_title = 12,
   size_axis_title = base_size,
@@ -27,6 +27,11 @@ plot.see_binned_residuals <- function(
     ltitle <- NULL
   }
   dots <- list(...)
+
+  # handle alias
+  if (!is.null(dots[["linewidth"]])) {
+    size_line <- dots[["linewidth"]]
+  }
 
   # set defaults
   term <- attr(x, "term", exact = TRUE)
@@ -72,7 +77,7 @@ plot.see_binned_residuals <- function(
         se = FALSE,
         formula = y ~ s(x, bs = "tp"),
         colour = colors[3],
-        linewidth = linewidth
+        linewidth = size_line
       )
   }
 
@@ -107,7 +112,7 @@ plot.see_binned_residuals <- function(
       ggplot2::geom_point(ggplot2::aes(y = .data$ybar), size = size_point) +
       ggplot2::geom_errorbar(
         ggplot2::aes(ymin = .data$CI_low, ymax = .data$CI_high),
-        linewidth = linewidth,
+        linewidth = size_line,
         width = 0
       )
   } else {
@@ -122,7 +127,7 @@ plot.see_binned_residuals <- function(
           ymax = .data$CI_high,
           colour = .data$group
         ),
-        linewidth = linewidth,
+        linewidth = size_line,
         width = 0
       )
   }