+++ R/plot.check_normality.R
@@ -7,7 +7,8 @@
 #'   Options are `"qq"` (default) for quantile-quantile (Q-Q) plots,
 #'   `"pp"` for probability-probability (P-P) plots, or
 #'   `"density"` for density overlay plots.
-#' @param linewidth Numeric value specifying size of line geoms.
+#' @param size_line Numeric value specifying size of line geoms. `linewidth` is
+#'   an alias for `size_line`.
 #' @param alpha_dot Numeric value specifying alpha level of the point geoms.
 #' @param alpha Numeric value specifying alpha level of the confidence bands.
 #' @param colors Character vector of length two, indicating the colors (in
@@ -44,7 +45,7 @@ plot.see_check_normality <- function(
   x,
   type = "qq",
   data = NULL,
-  linewidth = 0.8,
+  size_line = 0.8,
   size_point = 2,
   size_title = 12,
   size_axis_title = base_size,
@@ -65,6 +66,12 @@ plot.see_check_normality <- function(
     model <- data
   }
 
+  # handle alias
+  dots <- list(...)
+  if (!is.null(dots[["linewidth"]])) {
+    size_line <- dots[["linewidth"]]
+  }
+
   # for GLM, only halfnormal Q-Q plots
   if (inherits(model, "glm")) {
     type <- "qq"
@@ -76,7 +83,7 @@ plot.see_check_normality <- function(
       attributes(x)$re_qq,
       theme = theme,
       size_point = size_point,
-      linewidth = linewidth,
+      size_line = size_line,
       alpha_level = alpha,
       size_axis_title = size_axis_title,
       size_title = size_title,
@@ -88,7 +95,7 @@ plot.see_check_normality <- function(
     if (inherits(model, "performance_simres")) {
       return(plot.see_performance_simres(
         model,
-        linewidth = linewidth,
+        size_line = size_line,
         size_point = size_point,
         alpha = alpha,
         alpha_dot = alpha_dot,
@@ -106,7 +113,7 @@ plot.see_check_normality <- function(
     .plot_diag_qq(
       dat,
       size_point = size_point,
-      linewidth = linewidth,
+      size_line = size_line,
       size_axis_title = size_axis_title,
       size_title = size_title,
       base_size = base_size,
@@ -123,7 +130,7 @@ plot.see_check_normality <- function(
     dat <- .residuals_density(model)
     .plot_diag_norm(
       dat,
-      linewidth = linewidth,
+      size_line = size_line,
       alpha_level = alpha,
       base_size = base_size,
       size_axis_title = size_axis_title,
@@ -135,7 +142,7 @@ plot.see_check_normality <- function(
     .plot_diag_pp(
       dat,
       size_point = size_point,
-      linewidth = linewidth,
+      size_line = size_line,
       base_size = base_size,
       size_axis_title = size_axis_title,
       size_title = size_title,
@@ -233,7 +240,7 @@ plot.see_check_normality <- function(
 
 .plot_diag_norm <- function(
   x,
-  linewidth,
+  size_line,
   size_axis_title = 10,
   size_title = 12,
   alpha_level = 0.2,
@@ -261,7 +268,7 @@ plot.see_check_normality <- function(
     ggplot2::geom_line(
       mapping = ggplot2::aes(y = .data$curve),
       colour = colors[1],
-      linewidth = linewidth,
+      linewidth = size_line,
       na.rm = TRUE
     ) +
     ggplot2::labs(
@@ -280,7 +287,7 @@ plot.see_check_normality <- function(
 .plot_diag_qq <- function(
   x,
   size_point,
-  linewidth,
+  size_line,
   size_axis_title = 10,
   size_title = 12,
   alpha_level = 0.2,
@@ -332,7 +339,7 @@ plot.see_check_normality <- function(
       ),
       ggplot2::geom_qq_line(
         ggplot2::aes(sample = .data$y),
-        linewidth = linewidth,
+        linewidth = size_line,
         colour = colors[1],
         distribution = qhalfnorm,
         na.rm = TRUE
@@ -356,7 +363,7 @@ plot.see_check_normality <- function(
         detrend = detrend
       ),
       qqplotr::stat_qq_line(
-        linewidth = linewidth,
+        linewidth = size_line,
         colour = colors[1],
         detrend = detrend
       )
@@ -380,13 +387,13 @@ plot.see_check_normality <- function(
       if (detrend) {
         ggplot2::geom_hline(
           yintercept = 0,
-          linewidth = linewidth,
+          linewidth = size_line,
           colour = colors[1],
           na.rm = TRUE
         )
       } else {
         ggplot2::geom_qq_line(
-          linewidth = linewidth,
+          linewidth = size_line,
           colour = colors[1],
           na.rm = TRUE
         )
@@ -441,7 +448,7 @@ plot.see_check_normality <- function(
 .plot_diag_pp <- function(
   x,
   size_point,
-  linewidth,
+  size_line,
   size_axis_title = base_size,
   size_title = 12,
   alpha_level = 0.2,
@@ -468,7 +475,7 @@ plot.see_check_normality <- function(
         bandType = method
       ) +
       qqplotr::stat_pp_line(
-        linewidth = linewidth,
+        linewidth = size_line,
         colour = colors[1],
         detrend = detrend
       ) +
@@ -490,7 +497,7 @@ plot.see_check_normality <- function(
     p_plot <- ggplot2::ggplot(x, ggplot2::aes(x = .data$probs, y = .data$y)) +
       ggplot2::geom_abline(
         slope = if (detrend) 0 else 1,
-        linewidth = linewidth,
+        linewidth = size_line,
         colour = colors[1]
       ) +
       geom_point2(
@@ -526,7 +533,7 @@ plot.see_check_normality <- function(
 .plot_diag_reqq <- function(
   x,
   size_point,
-  linewidth,
+  size_line,
   size_axis_title = base_size,
   size_title = 12,
   panel = TRUE,
@@ -563,7 +570,7 @@ plot.see_check_normality <- function(
       ggplot2::stat_smooth(
         method = "lm",
         alpha = alpha_level,
-        linewidth = linewidth,
+        linewidth = size_line,
         formula = y ~ x,
         colour = colors[1]
       ) +