+++ R/plot.check_outliers.R
@@ -64,7 +64,7 @@
 plot.see_check_outliers <- function(
   x,
   size_text = 3.5,
-  linewidth = 0.8,
+  size_line = 0.8,
   size_title = 12,
   size_axis_title = base_size,
   base_size = 10,
@@ -81,6 +81,12 @@ plot.see_check_outliers <- function(
   # need to know the method first, because we change the default plot type
   # depending on the method
   outlier_methods <- attr(x, "method", exact = TRUE)
+
+  # handle alias
+  dots <- list(...)
+  if (!is.null(dots[["linewidth"]])) {
+    size_line <- dots[["linewidth"]]
+  }
 
   # validate that the method is correct
   if (length(outlier_methods) == 0) {
@@ -114,7 +120,7 @@ plot.see_check_outliers <- function(
       influential_obs,
       show_labels = show_labels,
       size_text = size_text,
-      linewidth = linewidth,
+      size_line = size_line,
       size_axis_title = size_axis_title,
       size_title = size_title,
       base_size = base_size,