+++ R/plot.check_predictions.R
@@ -116,7 +116,7 @@ data_plot.performance_pp_check <- function(x, type = "density", ...) {
 #' @export
 print.see_performance_pp_check <- function(
   x,
-  linewidth = 0.5,
+  size_line = 0.5,
   size_point = 2,
   size_bar = 0.7,
   size_axis_title = base_size,
@@ -132,9 +132,15 @@ print.see_performance_pp_check <- function(
   orig_x <- x
   check_range <- isTRUE(attributes(x)$check_range)
 
+  # handle alias
+  dots <- list(...)
+  if (!is.null(dots[["linewidth"]])) {
+    size_line <- dots[["linewidth"]]
+  }
+
   p1 <- .prepare_pp_check(
     x,
-    linewidth = linewidth,
+    size_line = size_line,
     size_point = size_point,
     size_bar = size_bar,
     size_axis_title = size_axis_title,
@@ -164,7 +170,7 @@ print.see_performance_pp_check <- function(
 #' @export
 plot.see_performance_pp_check <- function(
   x,
-  linewidth = 0.5,
+  size_line = 0.5,
   size_point = 2,
   size_bar = 0.7,
   size_axis_title = base_size,
@@ -181,9 +187,15 @@ plot.see_performance_pp_check <- function(
   x_limits <- .default_value(x, "x_limits", x_limits)
   type <- .default_value(x, "type", type)
 
+  # handle alias
+  dots <- list(...)
+  if (!is.null(dots[["linewidth"]])) {
+    size_line <- dots[["linewidth"]]
+  }
+
   p1 <- .prepare_pp_check(
     x,
-    linewidth = linewidth,
+    size_line = size_line,
     size_point = size_point,
     size_bar = size_bar,
     size_axis_title = size_axis_title,
@@ -211,7 +223,7 @@ plot.see_performance_pp_check <- function(
 # internal function, used by both the `print()` and `plot()` -----------------
 .prepare_pp_check <- function(
   x,
-  linewidth = 0.5,
+  size_line = 0.5,
   size_point = 2,
   size_bar = 0.7,
   size_axis_title = base_size,
@@ -261,7 +273,7 @@ plot.see_performance_pp_check <- function(
   # Build main plot
   .plot_pp_check(
     x,
-    linewidth = linewidth,
+    size_line = size_line,
     size_point = size_point,
     alpha_line = alpha_line,
     theme = theme,
@@ -279,7 +291,7 @@ plot.see_performance_pp_check <- function(
 
 .plot_pp_check <- function(
   x,
-  linewidth,
+  size_line,
   size_point,
   alpha_line,
   theme = NULL,
@@ -311,7 +323,7 @@ plot.see_performance_pp_check <- function(
       x,
       colors,
       info,
-      linewidth,
+      size_line,
       size_point,
       alpha_line,
       ...
@@ -342,7 +354,7 @@ plot.see_performance_pp_check <- function(
       x,
       colors,
       info,
-      linewidth,
+      size_line,
       size_point,
       alpha_line,
       type,
@@ -361,7 +373,7 @@ plot.see_performance_pp_check <- function(
       x,
       colors,
       info,
-      linewidth,
+      size_line,
       alpha_line,
       bandwidth,
       ...
@@ -397,7 +409,7 @@ plot.see_performance_pp_check <- function(
   x,
   colors,
   info,
-  linewidth,
+  size_line,
   alpha_line,
   bandwidth,
   ...
@@ -425,8 +437,8 @@ plot.see_performance_pp_check <- function(
     ) +
     ggplot2::scale_linewidth_manual(
       values = c(
-        "Observed data" = 1.7 * linewidth,
-        "Model-predicted data" = linewidth
+        "Observed data" = 1.7 * size_line,
+        "Model-predicted data" = size_line
       ),
       guide = "none"
     ) +
@@ -457,7 +469,7 @@ plot.see_performance_pp_check <- function(
   x,
   colors,
   info,
-  linewidth,
+  size_line,
   size_point,
   alpha_line,
   type = "discrete_dots",
@@ -531,7 +543,7 @@ plot.see_performance_pp_check <- function(
         ),
         position = ggplot2::position_nudge(x = 0.2),
         size = 0.4 * size_point,
-        linewidth = linewidth,
+        linewidth = size_line,
         stroke = 0,
         shape = 16
       ) +
@@ -632,7 +644,7 @@ plot.see_performance_pp_check <- function(
   x,
   colors,
   info,
-  linewidth,
+  size_line,
   size_point,
   alpha_line,
   ...
@@ -656,7 +668,7 @@ plot.see_performance_pp_check <- function(
       ),
       position = ggplot2::position_nudge(x = 0.2),
       size = 0.4 * size_point,
-      linewidth = linewidth,
+      linewidth = size_line,
       stroke = 0,
       shape = 16
     ) +